Move runtime scripts into trex subdirectory
