+++ runtime/trex.lua
@@ -4,17 +4,17 @@ print("Loading trex runtime from current working directory!")
 
 -- Test the custom require function by loading a module
 print("\n1. Loading utils module...")
-local utils = require("utils")
+local utils = require("trex/utils")
 utils.greet("Trex Runtime")
 
 -- Test relative imports work
 print("\n2. Testing relative module loading...")
-local utils2 = require("utils")  -- Should load from trex_runtime/utils.lua
+local utils2 = require("trex/utils")  -- Should load from trex_runtime/utils.lua
 utils2.log("Relative import test successful")
 
 -- Test advanced module with nested imports
 print("\n3. Loading advanced module (tests nested imports)...")
-local advanced = require("advanced")
+local advanced = require("trex/advanced")
 advanced.demonstrate_imports()
 
 -- Test terminal functions from modules
