+++ runtime/trex/utils.lua
